scripts/requirements.txt
# Requirements for tictactoe and tape-stacking demo scripts
# Install with: pip install -r scripts/requirements.txt

# Azure OpenAI (GPT-4.1 vision + Whisper transcription)
openai>=1.0.0

# Computer vision
opencv-python>=4.8.0
numpy>=1.24.0

# Audio capture + file I/O (tape_stack_play.py voice input)
sounddevice>=0.4.6
soundfile>=0.12.0

# Environment variable loading (.env file)
python-dotenv>=1.0.0

# Signal smoothing (smooth_episode.py)
scipy>=1.11.0
